adding new temporary sprites, adjusting levelConfig

+++ core/src/main/java/com/ainspiring/entities/FightingEntity.java
@@ -8,7 +8,7 @@ import com.badlogic.gdx.math.Vector2;
 public abstract class FightingEntity extends Entity {
     
     protected int damage;
-    protected int speed;
+    protected float speed;
     protected Animation<TextureRegion> runAnimation;
     protected Animation<TextureRegion> walkAnimation;
     protected Animation<TextureRegion> specialAttackAnimation;
@@ -16,7 +16,7 @@ public abstract class FightingEntity extends Entity {
     protected Animation<TextureRegion> hurtAnimation;
     protected Animation<TextureRegion> dieAnimation;
 
-    public FightingEntity(Texture image, Vector2 position, int health, int cost, int damage, int speed) {
+    public FightingEntity(Texture image, Vector2 position, int health, int cost, int damage, float speed) {
         super(image, position, health, cost);
         this.damage = damage;
         this.speed = speed;
